moved ABC classes from group.py to new module abc.py
